Fix: EmployeeManagement type error
The error "Object literal may only specify known properties, and 'userId' does not exist in type 'Omit<Employee, "id">'" indicates that the `userId` property is being included in the `newEmployee` object, but it shouldn't be because of the `Omit` type. This commit removes the `userId` property from the `newEmployee` object literal to resolve the type error.